+++ azure-communication-ui/chat/src/main/java/com/azure/android/communication/ui/chat/repository/MessageRepositoryWriter.kt
@@ -12,7 +12,7 @@ import com.azure.android.communication.ui.chat.models.MessageInfoModel
 // - pageRetrieved
 // - messageEdited
 // - messageDeleted
-internal interface MessageRepositoryMiddlewareInterface {
+internal interface MessageRepositoryWriter {
     fun addLocalMessage(messageInfoModel: MessageInfoModel)
     fun addPage(page: List<MessageInfoModel>)
     fun addServerMessage(message: MessageInfoModel)
