+++ azure-communication-ui/chat/src/test/java/com/azure/android/communication/ui/chat/redux/middleware/repository/MessageRepositoryMiddlewareUnitTest.kt
@@ -7,13 +7,13 @@ import com.azure.android.communication.ui.chat.models.MessageInfoModel
 import com.azure.android.communication.ui.chat.redux.AppStore
 import com.azure.android.communication.ui.chat.redux.action.ChatAction
 import com.azure.android.communication.ui.chat.redux.state.ReduxState
-import com.azure.android.communication.ui.chat.repository.MessageRepositoryMiddlewareInterface
+import com.azure.android.communication.ui.chat.repository.MessageRepositoryWriter
 import com.azure.android.communication.ui.chat.service.sdk.wrapper.ChatMessageType
 import org.junit.Assert
 import org.junit.Test
 import org.mockito.kotlin.mock
 
-internal class RepositoryMiddlewareUnitTest {
+internal class MessageRepositoryMiddlewareUnitTest {
 
     @Test
     fun messageRepositoryMiddleware_invoke_when_invokedWithAnyAction_then_invokeNext() {
@@ -28,10 +28,10 @@ internal class RepositoryMiddlewareUnitTest {
         val actionToDispatch = ChatAction.SendMessage(message)
         var nextReceivedAction: ChatAction? = null
 
-        val mockMessageRepository = mock<MessageRepositoryMiddlewareInterface> {}
+        val mockMessageRepository = mock<MessageRepositoryWriter> {}
 
         val messageRepositoryMiddlewareImplementation =
-            RepositoryMiddlewareImpl(
+            MessageRepositoryMiddlewareImpl(
                 mockMessageRepository
             )
